Send intermediate certificates
The old certomat (see branch "python") was correctly doing this,
but I forgot to bring over this change to the Go one. Use both
X3 and X4 certificates because that's how LetsEncrypt recommends
to do it.

Fixes the problem reported in US-3338.

Also: Remove Dockerfile, because we do not run this in docker anymore.\